DcTracker: Expedite data stall recovery
The current stall recovery process has a pending packet queue which
needs to fill up before a stall is determined. But that process is
also multi-step, and it is currently expecting that queue to fill
between steps.
Waiting for the unacknowledged packets to grow after running the
connection cleanup step looks like something that would never happen,
or would take a very long time to, so expedite the stall recovery after
it begins: if any single packet stalls and there's already a recovery
in progress, run the follow up step immediately.

Change-Id: Ib16b2c47b5c832a812c8bd2a3d940862741ed9db