GsmSS: Use per-slot resources for carrier-bound options
The decision of whether an MVNO operator is roaming or not comes from
an mcc/mnc-specific resource. Relying on the global systems resources
to provide the appropriate result for each of the slots on a multi-SIM
device is just wrong, since each SIM will have its own preferences and
arrays of compatible/blocked carriers.

Change-Id: If3b45a4a4702cf45bc1d965e26c8df7ef3ff783b